pyproject.toml
[project]
name = "coursework"
version = "0.1.0"
description = ""
authors = [
    {name = "neosome_college",email = "daniildaresin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flake8 (>=7.2.0,<8.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "datetime (>=5.5,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlrd (>=2.0.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
